feat: add method to abstraciton

+++ src/HotChocolate/Spatial/src/Data/Filters/Types/Operations/GeometryDistanceOperationType.cs
@@ -5,7 +5,7 @@ using static HotChocolate.Data.Spatial.Filters.SpatialFilterOperations;
 
 namespace HotChocolate.Data.Spatial.Filters
 {
-    public class GeometryDistanceOperationType : ComparableOperationFilterInput<float>
+    public class GeometryDistanceOperationType : ComparableOperationFilterInput<double>
     {
         protected override void Configure(IFilterInputTypeDescriptor descriptor)
         {
